Fix ctrl-c handling
