+++ Cargo.toml
@@ -261,6 +261,7 @@ uuid = { version = "1.2.1", features = ["v4"] }
 walkdir = "2.2.5"
 wasmbin = "0.6"
 webbrowser = "1.0.2"
+windows-sys = "0.59"
 xdg = "2.5"
 
 # Vendor the openssl we rely on, rather than depend on a
