+++ crates/cli/src/subcommands/mod.rs
@@ -14,6 +14,7 @@ pub mod publish;
 pub mod repl;
 pub mod server;
 pub mod sql;
+pub mod start;
 pub mod subscribe;
 pub mod upgrade;
 pub mod version;
